The AppMenu interface causes a 10sec delay if we attempt it in KDE
One of the qdbus interfaces times out, which can block the HUD entirely.

The KDE interface covers everything the AppMenu does.

Unfortunately, KDE's gtk menu proxy only fetches a shallow depth of
menu items in the first "fetch", so it spends a lot more time
"expanding" each submenu. It's faster to use the gtk menubar interface
in those cases. So we attempt to use it first.