Keep a reference to the rofi_process so we can kill it if need to
If we press Alt multiple times, it can queue up attempts to open the
menu. Before I fixed the dbus timeout bug, I was mashing Alt to try
and bring the menu up. When it finally timed out and used another
interface, then closed the menu, a queued action would bring up the
menu again, 4-5 times even.

So we'll need to terminate an active process on key press it seems.