Expiremental shortcut display.
It probably wont be merged since it messes with the search input.
Since it does, it makes fuzzy search useless since almost everything
will match Ctrl or Alt which isn't intended.

The only way to properly do this would be to replace rofi with something
else, which is lots of work since you then have to reimplement the
search code as well.