init rofi process before fetching menu items
This should drop less keypresses of a fast typer. The KDE dbus proxy
isn't always fast, and you don't want to drop keypresses while waiting
for the menu items to populate. Previously, the user would need to
wait for the menu to appear after the menu is populated. Firefox can
take a few seconds depending on how many bookmarks are in the menu.