Fix bug where menuitems would not activate if there was an icon.
The icon is stripped from the menu_result, so we need to build 2 dicts
where the first is the formatted label that is returned in menu_result
and the other is the iconlabel string we pass to rofi to display.

We no longer pass the icon if show_icons is false. It does not matter
if we do, but we might as well follow the same pattern as shortcuts.