Support older distros while enabling features by default (Issue #25)
Check `rofi -version` to see if it supports icons, and ignoring pango formatting for shortcuts.

Icons need at least rofi 1.5.3.

The commit for ignoring pango formatting is currently in rofi master,
so we need to check for > 1.5.4 as I do not know if the next version
of rofi will be 1.5.5 or 1.6.0.